refactor: extract pure data helpers to @volto-hydra/helpers package
Move 15 pure functions out of packages/hydra-js/hydra.src.js into a new
sibling workspace package packages/helpers/ so they can be imported
server-side (e.g. from an Astro render endpoint) without dragging in
the Bridge class, DOM listeners, or postMessage plumbing.

Moved (with their private support helpers - slateToText, textToSlate,
computePagingUI, hierarchicalSortByPosition, _getContainerFields,
_findChangedInBlock):

  expandListingBlocks, ploneFetchItems, contentPath,
  buildQuerystringSearchBody, calculatePaging, staticBlocks,
  convertFieldValue, getFieldTypeString,
  isSlateFieldType, isTextareaFieldType, isPlainStringFieldType,
  isTextEditableFieldType,
  deepEqual, formDataContentEqual, findChangedUnit

hydra.src.js re-exports all 15 names from @volto-hydra/helpers for
back-compat - existing imports from @volto-hydra/hydra-js keep working
unchanged. The two helpers that previously touched module-scope
references (bridgeInstance for expandListingBlocks, getAuthHeaders for
ploneFetchItems) now reach for window.__hydraBridge and use an
SSR-safe cookie reader inline, both gated by typeof window guards.

ListingBlock.astro: replaced the empty placeholder with a real SSR
fetch through expandListingBlocks + ploneFetchItems. The rendered
HTML now contains the actual resolved items (~94 for the docs
listing) instead of two empty marker divs; if the API call throws,
the legacy markers are emitted so the bridge test's "two elements
share the parent UID" contract still holds.

Tests:
  - jest (packages/hydra-js): 54/54 pass
  - doc-examples (astro): 20/20 pass
  - doc-examples (svelte): 20/20 pass (back-compat verified)

hydra.js bundle: 456.6kb (esbuild) - no warnings.
