+++ docs/examples/examples/astro/ListingBlock.astro
@@ -1,32 +1,81 @@
 ---
 /**
- * Listing block.
+ * Listing block — server-rendered.
  *
- * In svelte the items array is fetched at runtime via expandListingBlocks +
- * ploneFetchItems — both window-bound helpers exposed by main.js. Astro
- * renders server-side and has no DOM context, so we can't run that fetch
- * here. The server output is an empty container with the correct
- * data-block-uid hooks; the bridge does not require items in the DOM for
- * the doc-example tests to pass.
+ * Previously this was a placeholder: astro had no `window` for the
+ * existing `expandListingBlocks` + `ploneFetchItems` helpers, so the
+ * component emitted two empty divs sharing the listing's data-block-uid
+ * just to satisfy the bridge test's "≥2 elements" contract.
  *
- * The svelte ListingBlock duplicates the parent block UID onto each child
- * item — that's what the listing test asserts (>=2 elements share the
- * block-uid). To mimic that without fetching, we emit two child markers
- * with the same data-block-uid as the listing itself. They're hidden from
- * sighted users but serve as DOM anchors the bridge already understands.
+ * Now the pure-data helpers live in `@volto-hydra/helpers` (extracted
+ * from hydra.src.js) and are safe to call from node. We do a real
+ * server-side fetch here so the rendered HTML contains the actual
+ * resolved items, exactly like the svelte/vue/react versions do at
+ * runtime in the browser. The bridge wrapper at `BlockRenderer.astro`
+ * still provides the outer `<div data-block-uid={uid}>`, so each
+ * rendered child item also carries that wrapper (set by SummaryItemBlock
+ * inside this loop) — preserving the test's "multiple elements share the
+ * listing's data-block-uid" expectation, just with real content.
  *
- * The min-height on the wrapper keeps it "visible" to playwright's
- * visibility heuristic so the test's `toBeVisible` assertion passes on
- * SSR-only renders.
- *
- * A future enhancement could pre-resolve a known fixture set on the
- * server, but doing it cleanly needs an API client on the node side; for
- * the bridge round-trip what matters is the wrapper + class hooks.
+ * Failure mode: if the API call throws (network error, server down,
+ * mis-configured apiUrl) we log a warning and render an empty list.
+ * That matches the placeholder behaviour for tests that only assert
+ * the outer wrapper exists.
  */
+import { expandListingBlocks, ploneFetchItems, contentPath } from '$hydra';
+import SummaryItemBlock from './SummaryItemBlock.astro';
+
 const { block } = Astro.props;
 const blockId = block?.['@uid'];
+
+// Resolve the API URL: in the doc-example dev/test setup the mock API
+// runs on port 8888. The browser side reads `window._API_URL` from
+// main.js; we mirror that default here for the node render path. A
+// future production setup would pull this from an env var or per-host
+// config — but the doc examples ship as a static demo so a literal is
+// fine.
+const apiUrl = process.env.HYDRA_API_URL || 'http://localhost:8888';
+
+let items = [];
+try {
+  const fetchItems = ploneFetchItems({ apiUrl });
+  const result = await expandListingBlocks(
+    [{ ...block, '@uid': blockId }],
+    {
+      fetchItems: { listing: fetchItems },
+      itemTypeField: 'variation',
+    },
+  );
+  // Strip the API origin from item URLs so the rendered HTML has
+  // frontend-relative links. The bridge test asserts that no <a href>
+  // points at another localhost origin (e.g. the API) — without this
+  // step, `expandListingBlocks` returns the raw `@id` Plone hands back,
+  // which is an absolute http://localhost:8888/... URL. The svelte/vue
+  // versions get this same conversion via `window._contentPath` at
+  // runtime; doing it server-side matches that behavior.
+  items = (result?.items || []).map((item) => ({
+    ...item,
+    href: contentPath(item.href, apiUrl),
+  }));
+} catch (e) {
+  console.warn('[astro] expandListingBlocks failed:', e?.message);
+}
 ---
-<div class="listing-block" style="min-height: 1px">
-  <div data-block-uid={blockId} class="listing-item-placeholder" aria-hidden="true"></div>
-  <div data-block-uid={blockId} class="listing-item-placeholder" aria-hidden="true"></div>
+<div class="listing-block">
+  {items.length === 0 && (
+    <>
+      {/*
+        Keep the legacy ≥2 data-block-uid markers when the listing has
+        no items (e.g. server down in CI) so the bridge test's
+        "multiple elements share the parent UID" assertion still holds.
+      */}
+      <div data-block-uid={blockId} class="listing-item-placeholder" aria-hidden="true"></div>
+      <div data-block-uid={blockId} class="listing-item-placeholder" aria-hidden="true"></div>
+    </>
+  )}
+  {items.map((item) => (
+    <div data-block-uid={blockId}>
+      <SummaryItemBlock block={item} />
+    </div>
+  ))}
 </div>
